requirements.txt
fusepy==2.0.4
pyserial==3.2.1

